\name{banknote}
\alias{banknote}
\docType{data}

\title{Swiss Banknote data}

\description{ The data set contain 6 measures ond 100 genuine and counterfeit Swiss franc banknotes. }

\usage{data(banknote)}

\format{A data frame with the following variables:

\describe{
\item{Status}{the status of the banknote: genuine or counterfeit}
\item{Length}{Length of bill (mm)}
\item{Left}{Width of left edge (mm)}
\item{Right}{Width of right edge (mm)}
\item{Bottom}{Bottom margin width (mm)}
\item{Top}{Top margin width (mm)}
\item{Diagonal}{Length of diagonal (mm)}

}}
\references{
	Flury, B. and Riedwyl, H. (1988).  \emph{Multivariate Statistics: A practical approach}. London: Chapman & Hall, Tables 1.1 and 1.2, pp. 5-8
}
% \examples{}
\keyword{data sets}
